requirements.txt
anthropic==0.52.0
psycopg2-binary==2.9.10
pgvector==0.3.6
sentence-transformers==4.1.0
spacy==3.8.13
python-dotenv==1.1.0
